+++ crml/sylo/src/lib.rs
@@ -4,8 +4,6 @@
 extern crate serde;
 
 extern crate parity_codec as codec;
-#[macro_use]
-extern crate parity_codec_derive;
 extern crate sr_io as runtime_io;
 extern crate sr_std as rstd;
 extern crate substrate_primitives;
@@ -20,5 +18,6 @@ pub mod e2ee;
 pub mod groups;
 pub mod inbox;
 pub mod response;
+pub mod vault;
 
 pub use rstd::vec;