+++ runtime/src/lib.rs
@@ -9,12 +9,14 @@ extern crate srml_support;
 #[macro_use]
 extern crate runtime_primitives;
 
-use cennznet_primitives::{AccountId, AccountIndex, Balance, BlockNumber, Hash, Index, SessionKey, Signature};
+use cennznet_primitives::{
+	AccountId, AccountIndex, Balance, BlockNumber, CennznetExtrinsic, Hash, Index, SessionKey, Signature,
+};
 #[cfg(feature = "std")]
 use council::seats as council_seats;
 use council::{motions as council_motions, voting as council_voting};
 use grandpa::fg_primitives::{self, ScheduledChange};
-use parity_codec_derive::{Decode, Encode};
+use parity_codec::{Decode, Encode};
 use rstd::prelude::*;
 use runtime_primitives::generic;
 use runtime_primitives::traits::{BlakeTwo256, Block as BlockT, Convert, DigestFor, NumberFor, StaticLookup};
@@ -33,6 +35,8 @@ use substrate_primitives::OpaqueMetadata;
 use version::NativeVersion;
 use version::RuntimeVersion;
 
+use generic_asset::{RewardAssetCurrency, SpendingAssetCurrency, StakingAssetCurrency};
+
 pub use consensus::Call as ConsensusCall;
 #[cfg(any(feature = "std", test))]
 pub use runtime_primitives::BuildStorage;
@@ -40,22 +44,21 @@ pub use runtime_primitives::{Perbill, Permill};
 pub use srml_support::StorageValue;
 pub use timestamp::Call as TimestampCall;
 
+pub use cennz_x::{ExchangeAddressGenerator, FeeRate};
 pub use sylo::device as sylo_device;
 pub use sylo::e2ee as sylo_e2ee;
 pub use sylo::groups as sylo_groups;
 pub use sylo::inbox as sylo_inbox;
 pub use sylo::response as sylo_response;
-use crate::cennznet_extrinsic::CennznetExtrinsic;
-
-mod cennznet_extrinsic;
+pub use sylo::vault as sylo_vault;
 
 /// Runtime version.
 pub const VERSION: RuntimeVersion = RuntimeVersion {
 	spec_name: create_runtime_str!("cennznet"),
 	impl_name: create_runtime_str!("centrality-cennznet"),
 	authoring_version: 1,
-	spec_version: 13,
-	impl_version: 13,
+	spec_version: 14,
+	impl_version: 14,
 	apis: RUNTIME_API_VERSIONS,
 };
 
@@ -122,13 +125,13 @@ impl session::Trait for Runtime {
 }
 
 impl staking::Trait for Runtime {
-	type Currency = generic_asset::Module<Self>;
+	type Currency = RewardAssetCurrency<Self>;
 	type OnRewardMinted = Treasury;
 	type Event = Event;
 }
 
 impl democracy::Trait for Runtime {
-	type Currency = generic_asset::Module<Self>;
+	type Currency = StakingAssetCurrency<Self>;
 	type Proposal = Call;
 	type Event = Event;
 }
@@ -148,7 +151,7 @@ impl council::motions::Trait for Runtime {
 }
 
 impl treasury::Trait for Runtime {
-	type Currency = generic_asset::Module<Self>;
+	type Currency = StakingAssetCurrency<Self>;
 	type ApproveOrigin = council_motions::EnsureMembers<_4>;
 	type RejectOrigin = council_motions::EnsureMembers<_2>;
 	type Event = Event;
@@ -160,6 +163,7 @@ impl contract::Trait for Runtime {
 	type Gas = u64;
 	type DetermineContractAddress = contract::SimpleAddressDeterminator<Runtime>;
 	type ComputeDispatchFee = contract::DefaultDispatchFeeComputor<Runtime>;
+	type ChargeFee = fees::Module<Self>;
 }
 
 impl sudo::Trait for Runtime {
@@ -181,12 +185,15 @@ impl generic_asset::Trait for Runtime {
 
 impl fees::Trait for Runtime {
 	type Event = Event;
-	type TransferAsset = GenericAsset;
+	type TransferAsset = SpendingAssetCurrency<Self>;
+	type OnFeeCharged = ();
 }
 
+impl rewards::Trait for Runtime {}
+
 impl cennz_x::Trait for Runtime {
-	type AccountId = AccountId;
 	type Event = Event;
+	type ExchangeAddressGenerator = ExchangeAddressGenerator<Self>;
 }
 
 impl attestation::Trait for Runtime {
@@ -206,6 +213,7 @@ impl sylo::device::Trait for Runtime {
 }
 impl sylo::response::Trait for Runtime {}
 impl sylo::inbox::Trait for Runtime {}
+impl sylo::vault::Trait for Runtime {}
 
 construct_runtime!(
 	pub enum Runtime with Log(InternalLog: DigestItem<Hash, SessionKey>) where
@@ -216,6 +224,7 @@ construct_runtime!(
 		System: system::{default, Log(ChangesTrieRoot)},
 		Aura: aura::{Module, Inherent(Timestamp)},
 		Timestamp: timestamp::{Module, Call, Storage, Config<T>, Inherent},
+		GenericAsset: generic_asset::{Module, Call, Storage, Config<T>, Event<T>},
 		Consensus: consensus::{Module, Call, Storage, Config<T>, Log(AuthoritiesChange), Inherent},
 		Indices: indices,
 		Session: session,
@@ -230,15 +239,16 @@ construct_runtime!(
 		Contract: contract::{Module, Call, Storage, Config<T>, Event<T>},
 		Sudo: sudo,
 		Fees: fees::{Module, Storage, Config<T>, Event<T>},
+		Rewards: rewards::{Module, Storage, Config<T>},
 		Attestation: attestation::{Module, Call, Storage, Event<T>},
 		Doughnut: doughnut::{Module, Call, Event<T>},
 		SpotExchange: cennz_x::{Module, Call, Storage, Config<T>, Event<T>},
-		GenericAsset: generic_asset::{Module, Call, Storage, Config<T>, Event<T>},
 		SyloGroups: sylo_groups::{Module, Call, Storage},
 		SyloE2EE: sylo_e2ee::{Module, Call, Event<T>, Storage},
 		SyloDevice: sylo_device::{Module, Call, Event<T>, Storage},
 		SyloInbox: sylo_inbox::{Module, Call, Storage},
 		SyloResponse: sylo_response::{Module, Call, Storage},
+		SyloVault: sylo_vault::{Module, Call, Storage},
 	}
 );
 
@@ -252,8 +262,8 @@ pub type SignedBlock = generic::SignedBlock<Block>;
 /// BlockId type as expected by this runtime.
 pub type BlockId = generic::BlockId<Block>;
 /// Unchecked extrinsic type as expected by this runtime.
-pub type UncheckedExtrinsic = CennznetExtrinsic<AccountId, Address, Index, Call, Signature>;
-//pub type UncheckedExtrinsic = generic::UncheckedMortalCompactExtrinsic<Address, Index, Call, Signature>;
+pub type UncheckedExtrinsic = CennznetExtrinsic<Address, Index, Call, Signature>;
+
 /// Extrinsic type that has already been checked.
 pub type CheckedExtrinsic = generic::CheckedExtrinsic<AccountId, Index, Call>;
 /// Executive: handles dispatch to the various modules.